cAER: cleanup of linenoise line editing library wrt warnings.
git-svn-id: https://svn.code.sf.net/p/jaer/code/cAER/trunk@7424 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ utils/ext/linenoise/linenoise.h
@@ -54,7 +54,7 @@ void linenoiseAddCompletion(linenoiseCompletions *, const char *);
 
 char *linenoise(const char *prompt);
 int linenoiseHistoryAdd(const char *line);
-int linenoiseHistorySetMaxLen(int len);
+int linenoiseHistorySetMaxLen(size_t len);
 int linenoiseHistorySave(const char *filename);
 int linenoiseHistoryLoad(const char *filename);
 void linenoiseClearScreen(void);
